Allow the OAuth token to be nil
It isn't an error if there's no OAuth token; it just means there's no OAuth token, and if we're
writing to a Public repository it doesn't matter. On the other hand we should catch errors where
adding the comment did fail, and log the failure approrpiately. While we're here, the #perform
method should return an approrpriate success/failure tuple.